Move xml creation into separate class
